requirements.txt
fastapi
uvicorn
python-dotenv
requests
assemblyai
Jinja2
python-multipart
google-generativeai
websockets